requirements.txt
adbutils==0.11.0
allure-pytest==2.9.43
allure-python-commons==2.9.43
apkutils2==1.0.0
Appium-Python-Client==1.2.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cigam==0.0.3
colorama==0.4.1
coverage==4.5.3
cryptography==35.0.0
decorator==4.4.0
Deprecated==1.2.13
deprecation==2.0.6
Django==3.2.7
docker-py==1.10.6
docker-pycreds==0.4.0
docopt==0.6.2
enum34==1.1.6
facebook-wda==0.3.4
filelock==3.3.1
fire==0.1.3
h11==0.12.0
html-testRunner==1.2.1
humanize==0.5.1
idna==2.8
importlib-metadata==0.18
iniconfig==1.1.1
Jinja2==3.0.1
logzero==1.5.0
lxml==4.3.4
MarkupSafe==2.0.1
more-itertools==7.1.0
namedlist==1.7
outcome==1.1.0
packaging==20.9
Pillow==8.3.1
pluggy==0.12.0
progress==1.5
py==1.10.0
pycparser==2.20
pyelftools==0.27
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.0
pytest==6.2.4
pytest-cov==2.7.1
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-repeat==0.8.0
pytest-rerunfailures==10.1
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
retry==0.9.2
selenium==3.141.0
six==1.12.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.0.3
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
uiautomator2==2.16.8
urllib3==1.26.7
wcwidth==0.1.7
websocket-client==1.2.1
weditor==0.2.3
whichcraft==0.6.0
wrapt==1.13.2
wsproto==1.0.0
xmltodict==0.12.0
zipp==0.5.1
